+++ CHANGELOG.md
@@ -2492,3 +2492,4 @@
 - [2026-07-02 19:47] perf(cache): apply RepaintBoundary on animated cache widgets
 - [2026-07-02 19:50] test(perf): add unit tests for perf business logic
 - [2026-07-02 20:14] fix(cache): prevent double tap duplicate action in cache
+- [2026-07-02 20:44] fix(perf): fix memory leak by cancelling perf subscription